feat: reimplement plugin for prettier v3
